Adapt beast example to new PollData result api
Avoid adding message if write is still to be done.
The only message we want to send via a broadcast is an updated result.
As result_ref() should always point to the most up to date version.
We can avoid adding a broadcast if a write is still to be done.