Reformat presentation
This was made as direct critisim of the German right-wing politicial party Afd. It was meant to draw a parallel in policies and mindset to the Nazi party. Challenging the audience to remember the past. I never showed it in the actual presentation, out of fear it may be misunderstood. For the same reasons I'm removing it from here.